backend/requirement.txt
Django==1.8.4
django-filter==0.11.0
djangorestframework==3.2.4
Markdown==2.6.2
requests==2.7.0
wheel==0.24.0
boto3==1.1.4
